requirements.txt
pygame==1.9.6
pymunk==5.6.0
svg.path==4.0.2
